Add exit to terminate program, only compile when .js(t) or .blueprint files are changed

+++ lib/architecture-js/watcher.rb
@@ -27,6 +27,7 @@ module ArchitectureJS
 
       def update_files(files, action)
         files.each do |f|
+          next unless f.match(/\.jst?$|\.blueprint$/)
           f = File.basename f
           if action == "deleted"
             FileUtils.rm_rf("#{@project.root}/#{@project.config[:build_dir]}/#{f}") if File.exists? "#{@project.root}/#{@project.config[:build_dir]}/#{f}"
